formularios

+++ front_angular/src/app/components/pages/administrador/caller_number/sip/sip.component.ts
@@ -2,6 +2,8 @@ import { Component, OnInit, Input } from '@angular/core';
 import { Router, ActivatedRoute } from '@angular/router';
 import { SipService } from '@services/sip.service';
 
+import Swal from 'sweetalert2';
+
 @Component({
 	selector: 'sip',
 	templateUrl: './sip.component.html'
@@ -10,9 +12,7 @@ export class SipComponent implements OnInit {
 	@Input() Alias: string;
 	@Input() Numero: string;
 	@Input() Context: string;
-	@Input() Id: string;
-
-	public identy;
+	@Input() Id: number;
 
 	constructor(
 		private route: ActivatedRoute,
@@ -24,15 +24,25 @@ export class SipComponent implements OnInit {
 
 	eliminarsip() {
 		console.log(this.Id);
-		this.identy = this.route.snapshot.paramMap.get('id');
-		console.log(this.identy);
-
-		this.serviceSip.deleteSip(this.Id).subscribe(
-			rt => {
-				console.log(rt);
-			},
-			er => console.log(er),
-			() => console.log('terminado')
-		);
+		Swal.fire({
+			title: 'Esta seguro?',
+			text: 'El SIP se eliminará!',
+			icon: 'warning',
+			showCancelButton: true,
+			confirmButtonColor: '#3085d6',
+			cancelButtonColor: '#d33',
+			confirmButtonText: 'Sí, Eliminarlo!'
+		}).then(result => {
+			if (result.value) {
+				Swal.fire('Eliminado!');
+				this.serviceSip.deleteSip(this.Id).subscribe(
+					response => {
+						console.log(response);
+					},
+					er => console.log(er),
+					() => console.log('terminado')
+				);
+			}
+		});
 	}
 }
